+++ src/redux/actions/costperuse/get-cost-per-use-package-titles.js
@@ -1,10 +1,9 @@
 export const GET_COST_PER_USE_PACKAGE_TITLES = 'GET_COST_PER_USE_PACKAGE_TITLES';
 
-export const getCostPerUsePackageTitles = (packageId, filterData, loadMore) => ({
+export const getCostPerUsePackageTitles = (packageId, filterData) => ({
   type: GET_COST_PER_USE_PACKAGE_TITLES,
   payload: {
     id: packageId,
     filterData,
-    loadMore,
   },
 });